requirements.txt
requests>=2.25.0
Pillow>=9.0.0
numpy>=1.21.0
openexr>=1.7.0